layouts/default.html: support multiple galleries
Set up a second gallery of images and modify layouts/default.html to
support generating multiple lightboxes and handling multiple image
grids.

The post example is changed to show the two galleries.

The post variables no longer have a boolean indicating if a gallery is
to be shown, but instead provides a list of galleries to set up.